backend/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
flask-migrate
python-dotenv
PyJWT
Flask-CORS
